Created FileManager
Pulled load and save methods out of editor code and into own code for reuse between app sections
